Addressed comment by Nick.
Moved the actual test file DescribeEra.hs to cardano-ledger-test/src/Test/Cardano/Ledger/DescribeEra.hs
  Not sure how to run this test from there.
Moved Cardano/Crypto/VRF/Fake.hs to shelley/chain-and-ledger/executable-spec/src/Cardano/Crypto/VRF/Fake.hs
  This suports defining both StandardCrypto and TestCrypto in  Cardano.Ledger.Crypto
For each new Era XXX, XXX.Era has been rename to XXX.Self, e.g. Shelley.SheleyEra and Shelley.Self are the same type.
Removed the proxy type synonym.
