+++ alonzo/impl/src/Cardano/Ledger/DescribeEras.hs
@@ -20,19 +20,13 @@ module Cardano.Ledger.DescribeEras
   )
 where
 
-import Cardano.Crypto.DSIGN (Ed25519DSIGN, MockDSIGN)
-import Cardano.Crypto.Hash (Blake2b_224, Blake2b_256, MD5Prefix)
-import Cardano.Crypto.KES (MockKES, Sum6KES)
-import Cardano.Crypto.VRF.Praos
 import qualified Cardano.Ledger.Allegra as Allegra
 import qualified Cardano.Ledger.Alonzo as Alonzo
 import Cardano.Ledger.Core as Core
-import qualified Cardano.Ledger.Crypto as CryptoClass
+import Cardano.Ledger.Crypto (StandardCrypto, TestCrypto)
 import Cardano.Ledger.Era (WellFormed)
 import qualified Cardano.Ledger.Mary as Mary
 import qualified Cardano.Ledger.Shelley as Shelley
-import Shelley.Spec.Ledger.API (PraosCrypto)
-import Test.Cardano.Crypto.VRF.Fake (FakeVRF)
 
 type DescribesShelley era =
   ( WellFormed era,
@@ -82,45 +76,19 @@ type DescribesAlonzo era =
 class Checks era where
   checks :: Witness era -> Bool
 
-instance DescribesShelley (Shelley.Era c) => Checks (Shelley.Era c) where
+instance DescribesShelley (Shelley.Self c) => Checks (Shelley.Self c) where
   checks (Shelley _) = True
 
-instance DescribesAllegra (Allegra.Era c) => Checks (Allegra.Era c) where
+instance DescribesAllegra (Allegra.Self c) => Checks (Allegra.Self c) where
   checks (Allegra _) = True
 
-instance DescribesMary (Mary.Era c) => Checks (Mary.Era c) where
+instance DescribesMary (Mary.Self c) => Checks (Mary.Self c) where
   checks (Mary _) = True
 
-instance DescribesAlonzo (Alonzo.Era c) => Checks (Alonzo.Era c) where
+instance DescribesAlonzo (Alonzo.Self c) => Checks (Alonzo.Self c) where
   checks (Alonzo _) = True
 
-{------------------------------------------------------------------------------
- First construct concrete versions of Crypto where the Hashing
- is concrete. Without this we won't be able to Hash things
-------------------------------------------------------------------------------}
-data TestCrypto
-
-instance CryptoClass.Crypto TestCrypto where
-  type HASH TestCrypto = MD5Prefix 10
-  type ADDRHASH TestCrypto = MD5Prefix 8
-  type DSIGN TestCrypto = MockDSIGN
-  type KES TestCrypto = MockKES 10
-  type VRF TestCrypto = FakeVRF
-
-instance PraosCrypto TestCrypto
-
-data StandardCrypto
-
-instance CryptoClass.Crypto StandardCrypto where
-  type DSIGN StandardCrypto = Ed25519DSIGN
-  type KES StandardCrypto = Sum6KES Ed25519DSIGN Blake2b_256
-  type VRF StandardCrypto = PraosVRF
-  type HASH StandardCrypto = Blake2b_256
-  type ADDRHASH StandardCrypto = Blake2b_224
-
-instance PraosCrypto StandardCrypto
-
--- ===========================================
+-- ==========================================================
 
 -- | Evidence that a valid (predefined) crypto exists
 data Evidence c where
@@ -134,8 +102,8 @@ instance Show (Evidence c) where
 -- | Witness of a valid (predefined) era
 data Witness era where
   Shelley :: Evidence c -> Witness (Shelley.ShelleyEra c)
-  Mary :: Evidence c -> Witness (Mary.Era c)
-  Allegra :: Evidence c -> Witness (Allegra.Era c)
+  Mary :: Evidence c -> Witness (Mary.MaryEra c)
+  Allegra :: Evidence c -> Witness (Allegra.AllegraEra c)
   Alonzo :: Evidence c -> Witness (Alonzo.AlonzoEra c)
 
 instance Show (Witness e) where
