+++ alonzo/impl/test/test/Tests.hs
@@ -1,6 +1,5 @@
 module Main where
 
-import Test.Cardano.Ledger.Alonzo.DescribeEra (describeTest)
 import qualified Test.Cardano.Ledger.Alonzo.Serialisation.Tripping as Tripping
 import Test.Tasty
 
@@ -8,8 +7,7 @@ tests :: TestTree
 tests =
   testGroup
     "Alonzo tests"
-    [ Tripping.tests,
-      describeTest
+    [ Tripping.tests
     ]
 
 main :: IO ()
