+++ shelley-ma/impl/src/Cardano/Ledger/Allegra.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Allegra
   ( AllegraEra,
-    Era,
-    proxy,
+    Self,
     TxOut,
     Value,
     TxBody,
@@ -34,7 +34,6 @@ import Cardano.Ledger.ShelleyMA.TxBody ()
 import Cardano.Ledger.Val (Val ((<->)))
 import Data.Default.Class (def)
 import qualified Data.Map.Strict as Map
-import Data.Proxy (Proxy (..))
 import Shelley.Spec.Ledger.API hiding (PParams, Tx, TxBody, TxOut)
 import Shelley.Spec.Ledger.EpochBoundary (BlocksMade (..), emptySnapShots)
 
@@ -87,10 +86,7 @@ instance PraosCrypto c => ShelleyBasedEra (AllegraEra c)
 
 -- Self-Describing type synomyms
 
-proxy :: Proxy (ShelleyMAEra 'Allegra c)
-proxy = Proxy
-
-type Era c = ShelleyMAEra 'Allegra c
+type Self c = ShelleyMAEra 'Allegra c
 
 type Script era = Timelock (E.Crypto era)
