+++ shelley-ma/shelley-ma-test/src/Test/Cardano/Ledger/EraBuffet.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 -- | This module exports pieces one may use to make a new Era.
 --   The pieces come in 3 flavors:
@@ -37,43 +38,20 @@ module Test.Cardano.Ledger.EraBuffet
   )
 where
 
-import Cardano.Crypto.DSIGN (Ed25519DSIGN, MockDSIGN)
-import Cardano.Crypto.Hash (Blake2b_224, Blake2b_256, MD5Prefix)
-import Cardano.Crypto.KES (MockKES, Sum6KES)
-import Cardano.Crypto.VRF.Praos
-import Cardano.Ledger.Allegra (AllegraEra)
-import Cardano.Ledger.Core (Script, TxBody, Value, AuxiliaryData)
-import Cardano.Ledger.Crypto (HASH)
-import qualified Cardano.Ledger.Crypto as CryptoClass
 import Cardano.Ledger.Era (Crypto, Era)
-import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.Shelley (ShelleyEra)
+import Cardano.Ledger.Allegra (AllegraEra)
+import Cardano.Ledger.Mary (MaryEra)
 import Shelley.Spec.Ledger.API (PraosCrypto)
-import Test.Cardano.Crypto.VRF.Fake (FakeVRF)
+import Cardano.Ledger.Core (Script, TxBody, Value, AuxiliaryData)
+import Cardano.Ledger.Crypto(TestCrypto,StandardCrypto)
 
 {------------------------------------------------------------------------------
  First construct concrete versions of Crypto where the Hashing
- is concrete. Without this we won't be able to Hash things
+ is concrete, are imported from Cardano.Ledger.Crypto(TestCrypto,StandardCrypto)
 ------------------------------------------------------------------------------}
-data TestCrypto
-
-instance CryptoClass.Crypto TestCrypto where
-  type HASH TestCrypto = MD5Prefix 10
-  type ADDRHASH TestCrypto = MD5Prefix 8
-  type DSIGN TestCrypto = MockDSIGN
-  type KES TestCrypto = MockKES 10
-  type VRF TestCrypto = FakeVRF
 
 instance PraosCrypto TestCrypto
-
-data StandardCrypto
-
-instance CryptoClass.Crypto StandardCrypto where
-  type DSIGN StandardCrypto = Ed25519DSIGN
-  type KES StandardCrypto = Sum6KES Ed25519DSIGN Blake2b_256
-  type VRF StandardCrypto = PraosVRF
-  type HASH StandardCrypto = Blake2b_256
-  type ADDRHASH StandardCrypto = Blake2b_224
 
 instance PraosCrypto StandardCrypto
