+++ shelley/chain-and-ledger/shelley-spec-ledger-test/src/Test/Shelley/Spec/Ledger/ConcreteCryptoTypes.hs
@@ -21,7 +21,7 @@ import Cardano.Ledger.Crypto
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Shelley.Spec.Ledger.API (PraosCrypto)
 import Shelley.Spec.Ledger.BaseTypes (Seed)
-import Test.Cardano.Crypto.VRF.Fake (FakeVRF)
+import Cardano.Crypto.VRF.Fake (FakeVRF)
 
 -- | Mocking constraints used in generators
 type Mock c =
